hotfix

+++ kube_run.sh
@@ -1,17 +1,16 @@
 #!/bin/bash
 NS="default"
 REPOSITORY=$(echo $K8SREPOSITORY)
-
 SERVICES=$1
-USER=$3
 TASK=$2
+USER=$3
 
 if [ -z "$USER" ];
   then
     USER=$KUBEUSER
     echo USER=$USER
     export KUBECONFIG=~/minikube/$USER/config
-  else 
+  else
     echo USER=$USER
     export KUBECONFIG=~/minikube/$USER/config
 fi
@@ -21,20 +20,19 @@ cd $REPOSITORY
 echo $REPOSITORY
 
 #TODO
-#if [ $TASK=monolith" ];
-# then
-
-rm $REPOSITORY/devserver/tasks/$TASK.yaml
-if [ $? -ne 0 ];
+if [ $TASK != empty ]
   then
-    echo "old yaml file is deleted"
-  else
-    echo "new yaml file will be created"
+   rm $REPOSITORY/devserver/tasks/$TASK.yaml
+    if [ $? -ne 0 ];
+      then
+        echo "old yaml file is deleted"
+      else
+        echo "new yaml file will be created"
+    fi
+     for i in $SERVICES; do
+       make_tasks.sh $TASK $REPOSITORY $SERVICES
+     done;
 fi
-for i in $SERVICES; do
-  make_tasks.sh $TASK $REPOSITORY $SERVICES
-done;
-#fi
 
 echo devserver/bin/deploy.sh $NS "$SERVICES" $USER $TASK run
 devserver/bin/deploy.sh $NS "$SERVICES" $USER $TASK run
